+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Version 1.2-dev
+* Fixed some false positives related to yield usage in expression context
 * Fixed false positives for nested by-reference foreach loops wrapped in functions or class methods
 * Added detection of continue/break outside of loop/switch contexts
 * Added JSON output format